Updated card and sensor
Changed the main state of the template sensor to show if the current week is a raceweek. This is used by the lovelace card to show the card with countdown or upcoming events.

I also removed all the default values for the timestamp_custom as I understand its not needed. You might get an error in the logs if the calendar is unavailable, but I have not encountered that, so if you have those check the availability of the calendars.

Added the new sensor state '(no) race this week' to show the correct card when this week is race week.

Also added an extra couple lines of code to show the correct state if days until is below 1 week. You could use it, but its not recommended.